fixed warning and removed trailing whitespaces

+++ modules/video/src/lkpyramid.cpp
@@ -11,7 +11,7 @@
 //                For Open Source Computer Vision Library
 //
 // Copyright (C) 2000, Intel Corporation, all rights reserved.
-// Copyright (C) 2013, OpenCV Foundation, all rights reserved. 
+// Copyright (C) 2013, OpenCV Foundation, all rights reserved.
 // Third party copyrights are property of their respective owners.
 //
 // Redistribution and use in source and binary forms, with or without modification,
@@ -848,7 +848,7 @@ cv::Mat cv::estimateRigidTransform( InputArray src1, InputArray src2, bool fullA
     double scale = 1.;
     int i, j, k, k1;
 
-    RNG rng(-1);
+    RNG rng((uint64)-1);
     int good_count = 0;
 
     if( A.size() != B.size() )
